AudioContext: take the sample rate as a double
setSampleRate() took an int, so a host running at a fractional rate left
the context and the DSP class disagreeing -- AudioContext saw 1234 while
the plugin's own setSampleRate(double) saw 1234.5678, and processBlock
reads the rate back from the context. currentTime.sampleRate is already a
double and EventsToSignals::setSampleRate already takes one, so the
truncation bought nothing and only introduced the mismatch. It also meant a
sub-1Hz rate truncated to zero, making 1/sr infinite downstream.

clap-validator exercises 1234.5678, 12345.678, 45678.901 and 123456.78.

+++ source/app/MLAudioContext.cpp
@@ -111,7 +111,7 @@ AudioContext::AudioContext(size_t nInputs, size_t nOutputs)
   clear();
 }
 
-void AudioContext::setSampleRate(int r)
+void AudioContext::setSampleRate(double r)
 {
   currentTime.sampleRate = r;
   eventsToSignals.setSampleRate(r);
